requirements.txt
openai
requests
python-dotenv
gspread
google-auth
sqlalchemy
psycopg2-binary
Flask 